[access.fuse]: Add read_only option

+++ ufs/access/fuse.py
@@ -12,22 +12,28 @@ from fuse import LoggingMixIn, Operations, FuseOSError
 logger = logging.getLogger(__name__)
 
 class FUSEOps(LoggingMixIn, Operations):
-  def __init__(self, ufs: UFS) -> None:
+  def __init__(self, ufs: UFS, readonly = False) -> None:
     super().__init__()
     self._os = UOS(ufs)
     # self._lock = Lock()
+    self._readonly = readonly
 
-  def access(self, path, *args, **kwargs):
-    if not self._os.access(path, *args, **kwargs):
+  def access(self, path, amode):
+    if self._readonly and amode & os.W_OK:
+      raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), path)
+    if not self._os.access(path, amode):
       raise FuseOSError(errno.EACCES)
 
   def chmod(self, path, *args, **kwargs):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), path)
     return self._os.chmod(path, *args, **kwargs)
 
   def chown(self, path, *args, **kwargs):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), path)
     return self._os.chown(path, *args, **kwargs)
 
   def create(self, path, mode):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), path)
     return self._os.open(path, os.O_WRONLY | os.O_CREAT | os.O_TRUNC, mode)
 
   def flush(self, path, fh):
@@ -46,12 +52,15 @@ class FUSEOps(LoggingMixIn, Operations):
       'st_nlink', 'st_size', 'st_uid'))
 
   def link(self, target, source):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), source)
     self._os.link(target, source)
 
   def mkdir(self, path, *args, **kwargs):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), path)
     return self._os.mkdir(path, *args, **kwargs)
 
   def mknod(self, path, *args, **kwargs):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), path)
     return self._os.mknod(path, *args, **kwargs)
 
   def open(self, path, *args, **kwargs):
@@ -61,25 +70,21 @@ class FUSEOps(LoggingMixIn, Operations):
     return self._os.readlink(path, *args, **kwargs)
 
   def rmdir(self, path, *args, **kwargs):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), path)
     return self._os.rmdir(path, *args, **kwargs)
 
   def unlink(self, path, *args, **kwargs):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), path)
     return self._os.unlink(path, *args, **kwargs)
 
   def utimens(self, path, *args, **kwargs):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), path)
     return self._os.utime(path, *args, **kwargs)
 
   def read(self, path, size, offset, fh):
     # with self._lock:
-    logger.debug(f"-> read {path=}, {size=}, {offset=}, {fh=}")
     self._os.lseek(fh, offset, 0)
-    try:
-      result = self._os.read(fh, size)
-    except:
-      import traceback
-      logger.error(f"<- read {traceback.format_exc()}")
-      raise
-    logger.debug(f"<- read {result}")
+    result = self._os.read(fh, size)
     return result
 
   def readdir(self, path, fh):
@@ -89,6 +94,7 @@ class FUSEOps(LoggingMixIn, Operations):
     return self._os.close(fh)
 
   def rename(self, old, new):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), new)
     return self._os.rename(old, new)
 
   def statfs(self, path):
@@ -104,31 +110,25 @@ class FUSEOps(LoggingMixIn, Operations):
     return self._os.symlink(source, target)
 
   def truncate(self, path, length, fh=None):
+    if self._readonly: raise PermissionError(errno.EPERM, os.strerror(errno.EPERM), path)
     self._os.truncate(path, length)
 
   def write(self, path, data, offset, fh):
     # with self._lock:
-    logger.debug(f"-> write {path=}, {data=}, {offset=}, {fh=}")
-    try:
-      self._os.lseek(fh, offset, 0)
-      result = self._os.write(fh, data)
-      logger.debug(f"<- write {result}")
-    except:
-      import traceback
-      logger.error(f"<- write {traceback.format_exc()}")
-      raise
+    self._os.lseek(fh, offset, 0)
+    result = self._os.write(fh, data)
     return result
 
   getxattr = None
   listxattr = None
 
-def fuse(ufs_spec: dict, mount_dir: str):
+def fuse(ufs_spec: dict, mount_dir: str, readonly: bool):
   from fuse import FUSE
   with UFS.from_dict(**ufs_spec) as ufs:
-    FUSE(FUSEOps(ufs), mount_dir, nothreads=True, foreground=True)
+    FUSE(FUSEOps(ufs, readonly=readonly), mount_dir, nothreads=True, foreground=True)
 
 @contextlib.contextmanager
-def fuse_mount(ufs: UFS, mount_dir: str = None):
+def fuse_mount(ufs: UFS, mount_dir: str = None, readonly: bool = False):
   import signal
   import tempfile
   import functools
@@ -138,7 +138,7 @@ def fuse_mount(ufs: UFS, mount_dir: str = None):
   mp_spawn = mp.get_context('spawn')
   mount_dir_resolved = pathlib.Path(mount_dir or tempfile.mkdtemp())
   try:
-    with active_process(mp_spawn.Process(target=fuse, args=(ufs.to_dict(), str(mount_dir_resolved))), terminate_signal=signal.SIGINT):
+    with active_process(mp_spawn.Process(target=fuse, args=(ufs.to_dict(), str(mount_dir_resolved), readonly)), terminate_signal=signal.SIGINT):
       wait_for(functools.partial(safe_predicate, mount_dir_resolved.is_mount))
       yield mount_dir_resolved
   finally:
@@ -150,5 +150,5 @@ if __name__ == '__main__':
   ufs = UFS.from_dict(**json.loads(os.environ.pop('UFS_SPEC')))
   mount_dir = pathlib.Path(sys.argv[1])
   assert mount_dir.exists()
-  with fuse_mount(ufs, mount_dir):
+  with fuse_mount(ufs, mount_dir, bool(os.environ.pop('UFS_READONLY', ''))):
     threading.Event().wait()
